@@ -1,8 +1,8 @@
 package com.thomas_sell.torcs2;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentActivity;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentActivity;
 import android.view.InflateException;
 import android.view.LayoutInflater;
 import android.view.View;
